Removed the use of std::vector
